refactor: remove ai prefix from remediation proposal worker
