+++ Makefile
@@ -1,4 +1,4 @@
-TARGETS = access-analyzer admin-checker cloudsploit guard-duty portscan ai-remediation-proposal
+TARGETS = access-analyzer admin-checker cloudsploit guard-duty portscan remediation-proposal
 BUILD_TARGETS = $(TARGETS:=.build)
 BUILD_CI_TARGETS = $(TARGETS:=.build-ci)
 IMAGE_PUSH_TARGETS = $(TARGETS:=.push-image)
@@ -103,11 +103,11 @@ enqueue-portscan:
 		--queue-url http://localhost:9324/queue/aws-portscan \
 		--message-body '{"aws_id":1001, "aws_data_source_id":1005, "data_source":"aws:portscan", "project_id":1001, "account_id":"315855282677", "assume_role_arn":"arn:aws:iam::315855282677:role/stg-security-monitor", "external_id":""}'
 
-.PHONY: enqueue-ai-remediation-proposal
-enqueue-ai-remediation-proposal:
+.PHONY: enqueue-remediation-proposal
+enqueue-remediation-proposal:
 	aws sqs send-message \
 		--endpoint-url http://localhost:9324 \
-		--queue-url http://localhost:9324/queue/aws-ai-remediation-proposal \
+		--queue-url http://localhost:9324/queue/aws-remediation-proposal \
 		--message-body '{"remediation_proposal_id":1001, "finding_id":2001, "project_id":1001, "data_source":"aws:cloudsploit", "aws_id":1001, "account_id":"123456789012", "assume_role_arn":"arn:aws:iam::123456789012:role/example", "external_id":""}'
 
 FAKE:
