+++ dockers/remediation-proposal/Dockerfile
@@ -4,10 +4,10 @@ COPY go.mod .
 COPY go.sum .
 RUN go mod download
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -buildvcs=false -o /go/bin/ai-remediation-proposal cmd/ai-remediation-proposal/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -buildvcs=false -o /go/bin/remediation-proposal cmd/remediation-proposal/main.go
 
 FROM public.ecr.aws/risken/base/risken-base:v0.0.1
-COPY --from=builder /go/bin/ai-remediation-proposal /usr/local/ai-remediation-proposal/bin/
+COPY --from=builder /go/bin/remediation-proposal /usr/local/remediation-proposal/bin/
 ENV DEBUG= \
     PROFILE_EXPORTER= \
     PROFILE_TYPES= \
@@ -16,11 +16,11 @@ ENV DEBUG= \
     AWS_SECRET_ACCESS_KEY= \
     AWS_SESSION_TOKEN= \
     SQS_ENDPOINT= \
-    AI_REMEDIATION_PROPOSAL_QUEUE_NAME= \
-    AI_REMEDIATION_PROPOSAL_QUEUE_URL= \
+    REMEDIATION_PROPOSAL_QUEUE_NAME= \
+    REMEDIATION_PROPOSAL_QUEUE_URL= \
     MAX_NUMBER_OF_MESSAGE= \
     WAIT_TIME_SECOND=20 \
     TZ=Asia/Tokyo
-WORKDIR /usr/local/ai-remediation-proposal
+WORKDIR /usr/local/remediation-proposal
 ENTRYPOINT ["/usr/local/bin/env-injector"]
-CMD ["bin/ai-remediation-proposal"]
+CMD ["bin/remediation-proposal"]
