+++ pkg/remediationproposal/message_test.go
@@ -1,4 +1,4 @@
-package airemediationproposal
+package remediationproposal
 
 import "testing"
